# apidefinition.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "0_86",
      "title": "IPS SERN API",
      "description": "OpenAPI 3.0 specification for the IPS SERN API."
    },
    "servers": [
      {
        "url": "https://ips-d2s-uksc-medsnomed-medsno.apps.ocp1.azure.dso.digital.mod.uk",
        "description": "Production Server"
      },
      {
      "url": "http://localhost:5049",
      "description": "Local Server"
    }
  ],
...
  "paths": {
    "/ips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create a new IPS record",
        "description": "Create a new IPS record.",
        "requestBody": {
          "description": "MongoDb JSON object representing the new IPS record.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON object of the created IPS record."
          }
        }
      }
    },
    "/ipsmany": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create multiple IPS records",
        "description": "Create multiple IPS records.",
        "requestBody": {
          "description": "Array of MongoDb JSON objects representing the new IPS records.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of MongoDb JSON objects of the created IPS records."
          }
        }
      }
    },
    "/ipsbundle": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create IPS records from an IPS Bundle",
        "description": "Create IPS records from an IPS Bundle.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/pushipsora": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Push IPS data to an external ORA system",
        "description": "Push IPS data to an external ORA system.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response from the ORA system."
          }
        }
      }
    },
    "/pushipsnld": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Push IPS data to an external NLD system",
        "description": "Push IPS data to an external NLD system.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response from the NLD system."
          }
        }
      }
    },
    "/ipsfrombeer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from BEER format",
        "description": "Create MongoDb IPS records from BEER format.",
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/ipsfromcda": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from CDA XML format",
        "description": "Create MongoDb IPS records from CDA XML format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/ipsfromhl72x": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Create MongoDb IPS records from HL7 2.x format",
        "description": "Create MongoDb IPS records from HL7 2.x format.",
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDb JSON as confirmation of record creation."
          }
        }
      }
    },
    "/convertmongo2beer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert MongoDB format to BEER format",
        "description": "Convert MongoDB format to BEER format.",
        "requestBody": {
          "description": "MongoDB - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertmongo2hl7": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert MongoDB format to HL7 2.3 format",
        "description": "Convert MongoDB format to HL7 2.3 format.",
        "requestBody": {
          "description": "MongoDB - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HL7 2.3 - Plain Text"
          }
        }
      }
    },
    "/convertbeer2mongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert BEER format to MongoDB format",
        "description": "Convert BEER format to MongoDB format.",
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/convertbeer2ips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert BEER format to IPS JSON format",
        "description": "Convert BEER format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "BEER - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertips2beer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert IPS JSON format to BEER format",
        "description": "Convert IPS JSON format to BEER format.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertips2plaintext": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert IPS JSON format to Plain Text Readable format",
        "description": "Convert IPS JSON format to Plain Text Readable format.",
        "requestBody": {
          "description": "IPS Bundle - JSON",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Human Readable - Plain Text"
          }
        }
      }
    },
    "/convertcdatoips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to IPS JSON format",
        "description": "Convert CDA XML format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertcdatobeer": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to BEER format",
        "description": "Convert CDA XML format to BEER format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BEER - Plain Text"
          }
        }
      }
    },
    "/convertcdatomongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert CDA XML format to MongoDB format",
        "description": "Convert CDA XML format to MongoDB format.",
        "requestBody": {
          "description": "CDA - XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/converthl72xtomongo": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert HL7 2.x format to MongoDB format",
        "description": "Convert HL7 2.x format to MongoDB format.",
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MongoDB JSON"
          }
        }
      }
    },
    "/converthl72xtoips": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert HL7 2.x format to IPS JSON format",
        "description": "Convert HL7 2.x format to IPS JSON format.",
        "parameters": [
          {
            "$ref": "#/components/parameters/XIPSFormat"
          }
        ],
        "requestBody": {
          "description": "HL7 2.x - Plain Text",
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPS Bundle JSON"
          }
        }
      }
    },
    "/convertxml": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Generic convert XML format to JSON format",
        "description": "Generic convert XML format to JSON format.",
        "requestBody": {
          "description": "XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON"
          }
        }
      }
    },
    "/convertfhirxml": {
      "post": {
        "tags": [
          "IPS"
        ],
        "summary": "Convert FHIR XML format to FHIR JSON format",
        "description": "Convert FHIR XML format to FHIR JSON format.",
        "requestBody": {
          "description": "FHIR XML",
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/xml": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FHIR JSON"
          }
        }
      }
    },
    "/npsVal": {
      "post": {
        "tags": [
          "Schema"
        ],
        "summary": "Validates an NPS format record.",
        "description": "Validates an NPS format record.",
        "requestBody": {
          "description": "MongoDb JSON object representing the NPS format record.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation report for NPS validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/NpsValidationReport"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/NpsValidationReport"
                }
              }
            }
          }
        }
      }
    },
    "/ipsNhsScrVal": {
      "post": {
        "tags": [
          "Schema"
        ],
        "summary": "Validate NHS SCR IPS FHIR JSON against schema",
        "description": "Validates a pasted FHIR JSON payload (Bundle or single resource) against the NHS SCR IPS (custom) schema and against the global FHIR R4 schema.",
        "requestBody": {
          "description": "FHIR JSON payload (Bundle or single resource).",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation report (custom schema + FHIR R4 structural validation).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/NhsScrValidationReport"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. missing/unknown resourceType).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/NhsScrValidationReport"
                }
              }
            }
          }
        }
      }
    },
    "/epsVal": {
      "post": {
        "tags": [
          "Schema"
        ],
        "summary": "Validate EPS IPS FHIR JSON against schema",
        "description": "Validates a pasted FHIR JSON payload (Bundle or single resource) against the EPS IPS (custom) schema and against the global FHIR R4 schema.",
        "requestBody": {
          "description": "FHIR JSON payload (Bundle or single resource).",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation report (custom schema + FHIR R4 structural validation).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/EpsValidationReport"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. missing/unknown resourceType).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/valschemas/EpsValidationReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/ipsraw/{id}": {
      "get": {
        "tags": [
          "IPS"
        ],
        "summary": "Retrieve raw IPS record by ID",
        "description": "Retrieve raw IPS record by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IPS record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object of the raw IPS record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "valschemas": {
...
      "BaseValidationReport": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Overall validation result."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/valschemas/ValidationIssue"
            },
            "description": "Primary validation errors."
          },
          "validationMode": {
            "type": "string",
            "description": "Validation mode used by the endpoint.",
            "example": "lenient"
          }
        },
        "required": [
          "valid",
          "errors"
        ]
      },
...
      "NpsValidationReport": {
        "allOf": [
          {
            "$ref": "#/components/valschemas/BaseValidationReport"
          },
          {
            "type": "object",
            "properties": {
              "validNps": {
                "type": "boolean"
              },
              "errorsNps": {
                "type": "array",
                "items": {
                  "$ref": "#/components/valschemas/ValidationIssue"
                }
              }
            },
            "required": [
              "validNps",
              "errorsNps"
            ]
          }
        ]
      }
    },
    "parameters": {
      "XIPSFormat": {
        "name": "x-ips-format",
        "in": "header",
        "description": "Optional format override. Available values: 'unified' (default), 'inter', 'nhsscr', 'euro', 'legacy'.",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
  }
}